apps/email-agent/pyproject.toml
[tool.poetry]
name = "email-agent"
version = "0.1.0"
description = "Email agent application"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
gunicorn = "^21.2.0"
core = {path = "../../packages/core", develop = true}
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
sqlmodel = "^0.0.14"
redis = "^5.0.0"
celery = "^5.3.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.25.0"
black = "^23.11.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

